move out test_splitk_reference
